move components package
