+++ component/component.go
