+++ filter/filter.go
@@ -1,6 +1,6 @@
 package filter
 
-import "github.com/yohamta/donburi/internal/component"
+import "github.com/yohamta/donburi/component"
 
 // LayoutFilter is a filter that filters entities based on their components.
 type LayoutFilter interface {
