+++ internal/storage/layout_test.go
@@ -3,11 +3,11 @@ package storage
 import (
 	"testing"
 
-	"github.com/yohamta/donburi/internal/component"
+	"github.com/yohamta/donburi/component"
 )
 
 func TestLayout(t *testing.T) {
-	compType := component.NewMockComponentType(struct{}{}, nil)
+	compType := NewMockComponentType(struct{}{}, nil)
 	components := []component.IComponentType{compType}
 	layout := NewLayout(components)
